requirements.txt
aiohttp==3.4.4
async-timeout==3.0.1
attrs==18.2.0
certifi==2018.8.24
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.2
django-extensions==2.1.3
django-filter==2.0.0
django-rest-swagger==2.2.0
djangorestframework==3.8.2
gunicorn==19.9.0
idna==2.7
idna-ssl==1.1.0
itypes==1.1.0
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.0
multidict==4.4.2
openapi-codec==1.3.2
psycopg2==2.7.5
psycopg2-binary==2.7.5
pytz==2018.5
requests==2.19.1
simplejson==3.16.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
yarl==1.2.6
